test: fix relocated-test imports and ADR paths (fixup for §2.22)
The prior commit staged only the git-mv renames — the `git add` aborted on a
stale (now-deleted) pathspec, so the conftest import-path updates and the ADR
reference updates were left uncommitted, leaving the relocated tests importing
the nonexistent tests.unit.core.explorer_pipeline.conftest.

- Point the 7 intra-suite imports at tests.unit.scripts.explorer_pipeline.conftest.
- Update the two adr-explorer-cli-surface.md references to the new test paths.
