+++ tests/unit/scripts/explorer_pipeline/test_duckdb_browser_contract.py
@@ -14,7 +14,7 @@ import pytest
 
 from _project.scripts.explorer_pipeline.contract import EXPLORER_READ_MODEL_VERSION
 from _project.scripts.explorer_pipeline.pipeline import SUBMISSION_MANIFEST_SUFFIX, ExplorerPipeline
-from tests.unit.core.explorer_pipeline.conftest import MINIMAL_BUNDLE
+from tests.unit.scripts.explorer_pipeline.conftest import MINIMAL_BUNDLE
 
 pytestmark = [pytest.mark.unit, pytest.mark.fast]
