+++ tests/unit/scripts/explorer_pipeline/test_legacy_artifact_guard.py
@@ -23,7 +23,7 @@ from pathlib import Path
 import pytest
 
 from _project.scripts.explorer_pipeline.pipeline import ExplorerPipeline
-from tests.unit.core.explorer_pipeline.conftest import MINIMAL_BUNDLE
+from tests.unit.scripts.explorer_pipeline.conftest import MINIMAL_BUNDLE
 
 pytestmark = [pytest.mark.unit, pytest.mark.fast]
