+++ tests/unit/scripts/explorer_pipeline/test_read_model_contract.py
@@ -37,7 +37,7 @@ from _project.scripts.explorer_pipeline.transformer import (
     BundleTransformer,
     _query_display_ms,
 )
-from tests.unit.core.explorer_pipeline.conftest import MINIMAL_BUNDLE
+from tests.unit.scripts.explorer_pipeline.conftest import MINIMAL_BUNDLE
 
 pytestmark = [pytest.mark.unit, pytest.mark.fast]
