+++ tests/unit/scripts/explorer_pipeline/test_visible_metrics_registry.py
@@ -25,7 +25,7 @@ import pytest
 import yaml
 
 from _project.scripts.explorer_pipeline.pipeline import ExplorerPipeline
-from tests.unit.core.explorer_pipeline.conftest import MINIMAL_BUNDLE
+from tests.unit.scripts.explorer_pipeline.conftest import MINIMAL_BUNDLE
 
 pytestmark = [pytest.mark.unit, pytest.mark.fast]
